+++ src/mako/cli/exceptions/CliException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\cli\exceptions;
+
+use RuntimeException;
+
+/**
+ * CLI exception.
+ */
+class CliException extends RuntimeException
+{
+
+}